Reverted change to double, as it didn't make any sense in the first place

+++ phase2/IWithdrawable.java
@@ -1,3 +1,3 @@
 public interface IWithdrawable {
-    double withdraw(double amount, double balance);
+    double withdraw(int amount, double balance);
 }
